CHROMIUM: config: Build rmtfs and IPA into the kernel for Qualcomm
Getting the modem started up takes a long time and we want to start it
ASAP.  Let's build the rmtfs related modules into the kernel.  This
gets things started earlier and there's not a huge amount of downside.

Doing this:
* Allows us to get rid of the "udev-trigger-rmtfs.conf" script that
  tried to start some of this stuff loading earlier.  Note that script
  wasn't quite right for two reasons: it triggered only one of the two
  needed things (it missed qcom-q6v5-mss) and it also didn't exclude
  the trigger from the later 'udev-trigger.conf', which means we were
  getting double-triggered.  Fixing the double-trigger is hard due to
  the way udev works.
* Allows us to take some stuff out of our init script that loaded
  these modules super early.

This adds about 50K to the size of the kernel and doesn't have any
measurable impact on the "pre-startup" boot perf metric.

NOTE: it's not important (for boot speed) to build IPA into the kernel
and we'd only eat 14K of kernel space if we took that out.  However,
as per b:168157047, there's a bug if IPA is loaded much later than the
rmtfs stuff so we add it in here too.

BUG=b:162465221, b:168157047
TEST=w/ all patches, WiFi starts a bit earlier (.7 seconds)

Cq-Depend: chromium:2545289
Change-Id: Idf2c8338c443b98a3f9985a544656b5404927949
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2545233
Reviewed-by: Evan Green <evgreen@chromium.org>
Reviewed-by: Abhishek Kumar <kuabhs@chromium.org>
(cherry picked from commit 6c37a7f134c7e4b79d9f6a03e447574bf59e533f)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2550175