UPSTREAM: drm/mcde: Fix RGB/BGR bug
I was confused when the graphics came out with blue
penguins on the DPI panel.

It turns out that the so-called "packed RGB666" mode
on the DSI formatter is incorrect: this mode is the
actual RGB888 mode, and the mode called RGB888 is
BGR888.

The claims that the MCDE had inverse RGB/BGR buffer
formats was wrong, so correct this and the buggy
register and everything is much more consistent, and
graphics look good on all targets, both DPI and
DSI.

Signed-off-by: Linus Walleij <linus.walleij@linaro.org>
Reviewed-by: Sam Ravnborg <sam@ravnborg.org>
Cc: phone-devel@vger.kernel.org
Cc: Stephan Gerhold <stephan@gerhold.net>
Link: https://patchwork.freedesktop.org/patch/msgid/20201117175413.869871-1-linus.walleij@linaro.org
(cherry picked from commit 77f512bde99ad1ebc88f094d18702fa9589c2206)
Signed-off-by: Sean Paul <seanpaul@chromium.org>

BUG=b:177892854
TEST=Tested on volteer, build-tested trogdor

Change-Id: I3d183d47a6dbaf62a45af0b253a41b3de6f6405b