FROMLIST: sched: Basic tracking of matching tasks
Introduce task_struct::core_cookie as an opaque identifier for core
scheduling. When enabled; core scheduling will only allow matching
task to be on the core; where idle matches everything.

When task_struct::core_cookie is set (and core scheduling is enabled)
these tasks are indexed in a second RB-tree, first on cookie value
then on scheduling function, such that matching task selection always
finds the most elegible match.

NOTE: *shudder* at the overhead...

NOTE: *sigh*, a 3rd copy of the scheduling function; the alternative
is per class tracking of cookies and that just duplicates a lot of
stuff for no raisin (the 2nd copy lives in the rt-mutex PI code).

BUG=b:152605392
TEST=run power_VideoCall test

(am from
https://lore.kernel.org/lkml/20210422123308.496975854@infradead.org/)

[Joel: folded fixes]
Signed-off-by: Joel Fernandes (Google) <joel@joelfernandes.org>
Signed-off-by: Peter Zijlstra (Intel) <peterz@infradead.org>
Tested-by: Don Hiatt <dhiatt@digitalocean.com>
Signed-off-by: Joel Fernandes <joelaf@google.com>
Change-Id: Iac3e640ee9a5a6362eb0a95316ee70b14a313403
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2880784
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>