FROMLIST: sched/fair: Fix forced idle sibling starvation corner case
If there is only one long running local task and the sibling is
forced idle, it  might not get a chance to run until a schedule
event happens on any cpu in the core.

So we check for this condition during a tick to see if a sibling
is starved and then give it a chance to schedule.

(am from
https://lore.kernel.org/lkml/20210422123308.617407840@infradead.org/)

BUG=b:152605392
TEST=run power_VideoCall test

Signed-off-by: Vineeth Pillai <viremana@linux.microsoft.com>
Signed-off-by: Peter Zijlstra (Intel) <peterz@infradead.org>
Tested-by: Don Hiatt <dhiatt@digitalocean.com>
Signed-off-by: Joel Fernandes <joelaf@google.com>
Change-Id: I2ad3366e5a1e517646f65a7c3b54267d19173b15
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2880786
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>