FROMGIT: remoteproc: sysmon: Ensure remote notification ordering
The reliance on the remoteproc's state for determining when to send
sysmon notifications to a remote processor is racy with regard to
concurrent remoteproc operations.

Further more the advertisement of the state of other remote processor to
a newly started remote processor might not only send the wrong state,
but might result in a stream of state changes that are out of order.

Address this by introducing state tracking within the sysmon instances
themselves and extend the locking to ensure that the notifications are
consistent with this state.

Fixes: 1f36ab3f6e3b ("remoteproc: sysmon: Inform current rproc about all active rprocs")
Fixes: 1877f54f75ad ("remoteproc: sysmon: Add notifications for events")
Fixes: 1fb82ee806d1 ("remoteproc: qcom: Introduce sysmon")
Cc: stable@vger.kernel.org
Reviewed-by: Rishabh Bhatnagar <rishabhb@codeaurora.org>
Link: https://lore.kernel.org/r/20201122054135.802935-2-bjorn.andersson@linaro.org
Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>
(cherry picked from commit 138a6428ba9023ae29e103e87a223575fbc3d2b7
 https://git.kernel.org/pub/scm/linux/kernel/git/andersson/remoteproc.git/ rproc-next)

BUG=b:149924251
TEST=Boot trogdor modem

Signed-off-by: Evan Green <evgreen@chromium.org>
Change-Id: I21d900ce17caf14eeaf25dbe76ce17dc0c3c78ca
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2566437
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Sujit Kautkar <sujitka@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2775847
Tested-by: Sujit Kautkar <sujitka@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Commit-Queue: Sujit Kautkar <sujitka@chromium.org>