UPSTREAM: drm/mcde: fix masking and bitwise-or on variable val
The masking of val with ~MCDE_CRX1_CLKSEL_MASK is currently being
ignored because there seems to be a missing bitwise-or of val in the
following statement.  Fix this by replacing the assignment of val
with a bitwise-or.

Fixes: d795fd322063 ("drm/mcde: Support DPI output")
Signed-off-by: Colin Ian King <colin.king@canonical.com>
Addresses-Coverity: ("Unused valued")
Signed-off-by: Linus Walleij <linus.walleij@linaro.org>
Link: https://patchwork.freedesktop.org/patch/msgid/20201124121528.395681-1-colin.king@canonical.com
(cherry picked from commit ab43108d0dd48d287e1b5bb9cd4a969457ed9b0e)
Signed-off-by: Sean Paul <seanpaul@chromium.org>

BUG=b:177892854
TEST=Tested on volteer, build-tested trogdor

Change-Id: I5cafafad1980d96808aeac64a623bfde42033677