UPSTREAM: drm: rcar-du: Skip encoder allocation for LVDS1 in dual-link mode
The rcar-du driver skips registration of the encoder for the LVDS1
output when LVDS is used in dual-link mode, as the LVDS0 and LVDS1 links
are bundled and handled through the LVDS0 output. It however still
allocates the encoder and immediately destroys it, which is pointless.
Skip allocation of the encoder altogether in that case.

Signed-off-by: Laurent Pinchart <laurent.pinchart+renesas@ideasonboard.com>
Reviewed-by: Jacopo Mondi <jacopo+renesas@jmondi.org>
Reviewed-by: Kieran Bingham <kieran.bingham+renesas@ideasonboard.com>
(cherry picked from commit 9a248605980a7547100d06ab9053f49f3ce43418)
Signed-off-by: Sean Paul <seanpaul@chromium.org>

BUG=b:187093397
TEST=Tested with single/multiple displays, suspend/resume, hotplugs on volteer/trogdor/dedede/zork, build tested on amd64/arm64-generic

Change-Id: I094f3b4c4168eb5a1044b8298d9c42a398d0e489