CHROMIUM: arm64: dts: qcom: Only enable camera nodes on CoachZ
We don't need the camera nodes in the trogdor.dtsi file as not all
trogdor boards use these nodes. Limit the import to be in CoachZ only so
that they're not default enabled in all the other trogdor boards.
Eventually this should be dropped entirely and the nodes should live in
sc7180.dtsi and then we can disable them in trogdor.dtsi and enable them
in coachz.dtsi but this is good enough for now.

BUG=b:173780970
TEST=emerge-strongbad chromeos-kernel-5_4

Signed-off-by: Stephen Boyd <swboyd@chromium.org>
Change-Id: Ibd8811ad918679cdb6545ebd17008ee742422367
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2585864
Reviewed-by: Tomasz Figa <tfiga@chromium.org>