CHROMIUM: soc: mediatek: Drop parent/child relationship for mfg domains on mt8173
Based on commit 4762f11e0c95 ("CHROMIUM: soc: mediatek: SCPSYS: Drop
parent/child relationship for mfg domains")

This change drops the genpd relationships between the various mfg power
domains and instead treats them as a single domains to be powered off/on
together. This also ignores errors when failing to power off any of the
mfg domains.

This is workaround for system freeze/hang that occurs when the mfg_2d
domains fails to power off, and should be reverted once the root cause
is found. This is the behaviour of 3.18 which experiences mfg_2d power
off failures, but the devices appear to continue functioning.

Cherry-picked from https://crrev.com/c/2589179

BUG=b:168839892 b:179950730
TEST=Build/deploy to hana. Run webgl aquarium, watch some youtube.

Signed-off-by: Anand K Mistry <amistry@google.com>
Signed-off-by: Eizan Miyamoto <eizan@chromium.org>
Cq-Depend: chromium:2689284
Change-Id: Ifd6680443845bc1bccdc52d11d37948be7225c85
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2689286
Reviewed-by: Anand K Mistry <amistry@chromium.org>