BACKPORT: FROMGIT: net: ipa: use state to determine channel command success
The result of issuing a channel control command should be that the
channel changes state.  If enabled, a completion interrupt signals
that the channel state has changed.  This interrupt is enabled by
gsi_channel_command() and disabled again after the command has
completed (or we time out).

There is a window of time--after the completion interrupt is disabled
but before the channel state is read--during which the command could
complete successfully without interrupting.  This would cause the
channel to transition to the desired new state.

So whether a channel command ends via completion interrupt or
timeout, we can consider the command successful if the channel
has entered the desired state (and a failure if it has not,
regardless of the cause).

Fixes: d6c9e3f506ae8 ("net: ipa: only enable generic command completion IRQ when needed");
Signed-off-by: Alex Elder <elder@linaro.org>
Signed-off-by: Jakub Kicinski <kuba@kernel.org>
(cherry picked from commit 6ffddf3b3d182d886d754cfafdf909ccb14f464b
 https://git.kernel.org/pub/scm/linux/kernel/git/netdev/net.git master)

Conflicts:
   drivers/net/ipa/gsi.c

Conflict resolved by removing newly added channel id from log message

BUG=b:175365709
TEST=Modem functional, no IPA hang

Signed-off-by: Sujit Kautkar <sujitka@chromium.org>
Change-Id: I15a5534137f132ec135bddc07f1791b0493f29a2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2606596
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2607195