FROMGIT: iommu/mediatek: Support up to 34bit iova in tlb flush
If the iova is 34bit, the iova[32][33] is the bit0/1 in the tlb flush
register. Add a new macro for this.

In the macro, since (iova + size - 1) may be end with 0xfff, then the
bit0/1 always is 1, thus add a mask.

Signed-off-by: Yong Wu <yong.wu@mediatek.com>
Reviewed-by: Tomasz Figa <tfiga@chromium.org>
Link: https://lore.kernel.org/r/20210111111914.22211-22-yong.wu@mediatek.com
Signed-off-by: Will Deacon <will@kernel.org>
(cherry picked from commit bfed873114c5fbb8982113abbab17cde7788dcba
 git://git.kernel.org/pub/scm/linux/kernel/git/will/linux.git for-joerg/mtk)

BUG=b:153618578
TEST=Boot asurada

Signed-off-by: Nicolas Boichat <drinkcat@chromium.org>
Change-Id: I04c7767d6b8c8245af171be8a151ab1e1d3ae151
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2670346
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Hsin-Yi Wang <hsinyi@chromium.org>