FROMGIT: soc: mediatek: pm-domains: Don't print an error if child domain is deferred
Child domains can be deferred by the core because one of its resources
is not available yet, in such case, it will print an error, but
later it will succeed to probe. Fix that using the dev_err_probe()
function so it only prints an error on a real error.

Signed-off-by: Enric Balletbo i Serra <enric.balletbo@collabora.com>
Reviewed-by: Hsin-Yi Wang <hsinyi@chromium.org>
Link: https://lore.kernel.org/r/20210113213012.67643-1-enric.balletbo@collabora.com
Signed-off-by: Matthias Brugger <matthias.bgg@gmail.com>

(cherry picked from commit dd65030295e20338bbb8238454c2e9546b6e5e17
https://git.kernel.org/pub/scm/linux/kernel/git/matthias.bgg/linux.git v5.11-next/soc)

BUG=b:178776793
TEST=boot kukui and check display

Signed-off-by: Hsin-Yi Wang <hsinyi@chromium.org>
Change-Id: I472cfd288049776de535340ed28b45b2c5954e72
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2662596
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>